refactor: provider into module
