+++ crates/provider/src/provider/mod.rs
@@ -1,0 +1,5 @@
+mod r#trait;
+pub use r#trait::{FilterPollerBuilder, Provider, SendableTx};
+
+mod root;
+pub use root::RootProvider;
